requirements.txt
glfw==2.1.0
numpy==1.20.1
PyOpenGL==3.1.5
pyrr==0.10.3
pyvista==0.29.0